+++ app/src/main/java/com/mapcode/map/InfoScreenPart.kt
@@ -2,10 +2,13 @@
 
 package com.mapcode.map
 
+import android.annotation.SuppressLint
+import android.os.Build
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.text.ClickableText
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.*
@@ -15,39 +18,45 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.mapcode.R
 import com.mapcode.theme.MapcodeTheme
-import kotlinx.coroutines.launch
 
 @Composable
 fun InfoArea(
     modifier: Modifier,
     viewModel: MapViewModel,
-    scaffoldState: ScaffoldState,
+    showSnackbar: (String) -> Unit,
     isVerticalLayout: Boolean
 ) {
     val uiState by viewModel.uiState.collectAsState()
-    val scope = rememberCoroutineScope()
     val copiedMessageStr = stringResource(R.string.copied_to_clipboard_snackbar_text)
     val onMapcodeClick = remember {
         {
             val copied = viewModel.copyMapcode()
-            if (copied) {
-                scope.launch {
-                    //dismiss current snack bar so they aren't queued up
-                    scaffoldState.snackbarHostState.currentSnackbarData?.dismiss()
-                    scaffoldState.snackbarHostState.showSnackbar(copiedMessageStr)
-                }
+            if (copied && Build.VERSION.SDK_INT <= Build.VERSION_CODES.S_V2) {
+                showSnackbar(copiedMessageStr)
+            }
+        }
+    }
+    val onLocationClick = remember {
+        {
+            viewModel.copyLocation()
+            if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.S_V2) {
+                showSnackbar(copiedMessageStr)
             }
         }
     }
@@ -56,12 +65,15 @@ fun InfoArea(
         modifier,
         uiState,
         onMapcodeClick = onMapcodeClick,
-        onAddressChange = viewModel::queryAddress,
+        onAddressChange = viewModel::onAddressTextChange,
+        onSubmitAddress = viewModel::onSubmitAddress,
         onTerritoryClick = viewModel::onTerritoryClick,
         onChangeLatitude = viewModel::onLatitudeTextChanged,
         onSubmitLatitude = viewModel::onSubmitLatitude,
+        onCopyLatitude = onLocationClick,
         onChangeLongitude = viewModel::onLongitudeTextChanged,
         onSubmitLongitude = viewModel::onSubmitLongitude,
+        onCopyLongitude = onLocationClick,
         isVerticalLayout = isVerticalLayout
     )
 }
@@ -71,10 +83,13 @@ private fun InfoArea(
     modifier: Modifier = Modifier,
     state: UiState,
     onAddressChange: (String) -> Unit = {},
+    onSubmitAddress: () -> Unit = {},
     onChangeLatitude: (String) -> Unit = {},
     onSubmitLatitude: () -> Unit = {},
+    onCopyLatitude: () -> Unit = {},
     onChangeLongitude: (String) -> Unit = {},
     onSubmitLongitude: () -> Unit = {},
+    onCopyLongitude: () -> Unit = {},
     onTerritoryClick: () -> Unit = {},
     onMapcodeClick: () -> Unit = {},
     isVerticalLayout: Boolean
@@ -84,10 +99,13 @@ private fun InfoArea(
             modifier,
             state,
             onAddressChange,
+            onSubmitAddress,
             onChangeLatitude,
             onSubmitLatitude,
+            onCopyLatitude,
             onChangeLongitude,
             onSubmitLongitude,
+            onCopyLongitude,
             onTerritoryClick,
             onMapcodeClick
         )
@@ -96,10 +114,13 @@ private fun InfoArea(
             modifier,
             state,
             onAddressChange,
+            onSubmitAddress,
             onChangeLatitude,
             onSubmitLatitude,
+            onCopyLatitude,
             onChangeLongitude,
             onSubmitLongitude,
+            onCopyLongitude,
             onTerritoryClick,
             onMapcodeClick
         )
@@ -114,6 +135,7 @@ private fun InfoAreaPreview() {
             mapcodeUi = MapcodeUi("AB.XY", "NLD", "Netherlands", 1, 1),
             addressUi = AddressUi(
                 "I am a very very very very very very extremely long address",
+                emptyList(),
                 AddressError.UnknownAddress("Street, City"),
                 AddressHelper.NoInternet,
             ),
@@ -128,20 +150,25 @@ private fun VerticalInfoArea(
     modifier: Modifier = Modifier,
     state: UiState,
     onAddressChange: (String) -> Unit,
+    onSubmitAddress: () -> Unit,
     onChangeLatitude: (String) -> Unit,
     onSubmitLatitude: () -> Unit,
+    onCopyLatitude: () -> Unit,
     onChangeLongitude: (String) -> Unit,
     onSubmitLongitude: () -> Unit,
+    onCopyLongitude: () -> Unit,
     onTerritoryClick: () -> Unit,
     onMapcodeClick: () -> Unit
 ) {
     Column(modifier) {
         AddressArea(
-            Modifier,
-            state.addressUi.address,
-            onAddressChange,
-            state.addressUi.helper,
-            state.addressUi.error
+            modifier = Modifier,
+            address = state.addressUi.address,
+            matchingAddresses = state.addressUi.matchingAddresses,
+            onChange = onAddressChange,
+            onSubmit = onSubmitAddress,
+            helper = state.addressUi.helper,
+            error = state.addressUi.error
         )
         Spacer(Modifier.height(8.dp))
         TerritoryBox(
@@ -168,6 +195,7 @@ private fun VerticalInfoArea(
             showInvalidError = state.locationUi.showLatitudeInvalidError,
             onSubmit = onSubmitLatitude,
             onChange = onChangeLatitude,
+            onCopy = onCopyLatitude
         )
         Spacer(Modifier.height(8.dp))
         LongitudeTextBox(
@@ -177,6 +205,7 @@ private fun VerticalInfoArea(
             showInvalidError = state.locationUi.showLongitudeInvalidError,
             onSubmit = onSubmitLongitude,
             onChange = onChangeLongitude,
+            onCopy = onCopyLongitude
         )
     }
 }
@@ -186,20 +215,25 @@ private fun HorizontalInfoArea(
     modifier: Modifier = Modifier,
     state: UiState,
     onAddressChange: (String) -> Unit,
+    onSubmitAddress: () -> Unit,
     onChangeLatitude: (String) -> Unit,
     onSubmitLatitude: () -> Unit,
+    onCopyLatitude: () -> Unit,
     onChangeLongitude: (String) -> Unit,
     onSubmitLongitude: () -> Unit,
+    onCopyLongitude: () -> Unit,
     onTerritoryClick: () -> Unit,
     onMapcodeClick: () -> Unit
 ) {
     Column(modifier) {
         AddressArea(
-            Modifier.fillMaxWidth(),
-            state.addressUi.address,
-            onAddressChange,
-            state.addressUi.helper,
-            state.addressUi.error
+            modifier = Modifier.fillMaxWidth(),
+            address = state.addressUi.address,
+            matchingAddresses = state.addressUi.matchingAddresses,
+            onChange = onAddressChange,
+            onSubmit = onSubmitAddress,
+            helper = state.addressUi.helper,
+            error = state.addressUi.error
         )
         Row(Modifier.padding(top = 8.dp)) {
             TerritoryBox(
@@ -232,6 +266,7 @@ private fun HorizontalInfoArea(
                 showInvalidError = state.locationUi.showLatitudeInvalidError,
                 onSubmit = onSubmitLatitude,
                 onChange = onChangeLatitude,
+                onCopy = onCopyLatitude
             )
             LongitudeTextBox(
                 modifier = Modifier
@@ -243,6 +278,7 @@ private fun HorizontalInfoArea(
                 showInvalidError = state.locationUi.showLongitudeInvalidError,
                 onSubmit = onSubmitLongitude,
                 onChange = onChangeLongitude,
+                onCopy = onCopyLongitude
             )
         }
     }
@@ -286,7 +322,8 @@ private fun LatitudeTextBox(
     placeHolder: String,
     showInvalidError: Boolean,
     onSubmit: () -> Unit,
-    onChange: (String) -> Unit
+    onChange: (String) -> Unit,
+    onCopy: () -> Unit
 ) {
     LatLngTextField(
         modifier = modifier,
@@ -296,7 +333,8 @@ private fun LatitudeTextBox(
         label = stringResource(R.string.latitude_text_field_label),
         clearButtonContentDescription = stringResource(R.string.clear_latitude_content_description),
         onSubmit = onSubmit,
-        onChange = onChange
+        onChange = onChange,
+        onCopy = onCopy
     )
 }
 
@@ -310,7 +348,8 @@ private fun LongitudeTextBox(
     placeHolder: String,
     showInvalidError: Boolean,
     onSubmit: () -> Unit,
-    onChange: (String) -> Unit
+    onChange: (String) -> Unit,
+    onCopy: () -> Unit
 ) {
     LatLngTextField(
         modifier = modifier,
@@ -320,10 +359,12 @@ private fun LongitudeTextBox(
         label = stringResource(R.string.longitude_text_field_label),
         clearButtonContentDescription = stringResource(R.string.clear_longitude_content_description),
         onSubmit = onSubmit,
-        onChange = onChange
+        onChange = onChange,
+        onCopy = onCopy
     )
 }
 
+@SuppressLint("UnrememberedMutableState")
 @Composable
 private fun LatLngTextField(
     modifier: Modifier = Modifier,
@@ -333,20 +374,48 @@ private fun LatLngTextField(
     label: String,
     clearButtonContentDescription: String,
     onSubmit: () -> Unit,
-    onChange: (String) -> Unit
+    onChange: (String) -> Unit,
+    onCopy: () -> Unit
 ) {
     Column(modifier) {
         val focusManager = LocalFocusManager.current
         val focusRequester = remember { FocusRequester() }
+        var textSelection: TextRange by remember { mutableStateOf(TextRange.Zero) }
+
+        val textValue: TextFieldValue by derivedStateOf { TextFieldValue(text, textSelection) }
+
+        var selectAllText: Boolean by remember { mutableStateOf(false) }
+
+        if (selectAllText) {
+            SideEffect {
+                textSelection = TextRange(0, text.length)
+                selectAllText = false
+            }
+        }
 
         OutlinedTextField(
             modifier = Modifier
                 .focusRequester(focusRequester)
+                .onFocusChanged { state ->
+                    if (state.isFocused) {
+                        selectAllText = true
+                    }
+                }
                 .fillMaxWidth(),
-            value = text,
+            value = textValue,
             singleLine = true,
-            label = { Text(label, maxLines = 1) },
-            onValueChange = onChange,
+            label = {
+                ClickableText(
+                    modifier = Modifier.testTag("latlngtextfield"),
+                    text = buildAnnotatedString { append(label) },
+                    maxLines = 1,
+                    onClick = { onCopy() }
+                )
+            },
+            onValueChange = { value ->
+                textSelection = value.selection
+                onChange(value.text)
+            },
             keyboardOptions = KeyboardOptions(
                 imeAction = ImeAction.Go,
                 keyboardType = KeyboardType.Decimal
@@ -409,7 +478,7 @@ private fun HeaderWithIcon(modifier: Modifier = Modifier, text: String, @Drawabl
 private fun MapcodeBox(
     modifier: Modifier = Modifier,
     code: String,
-    territory: String
+    territory: String?
 ) {
     Card(
         modifier = modifier,
@@ -426,10 +495,12 @@ private fun MapcodeBox(
                 MaterialTheme.typography.body1.copy(fontWeight = FontWeight.Bold).toSpanStyle()
 
             val styledString = buildAnnotatedString {
-                pushStyle(MaterialTheme.typography.body2.toSpanStyle())
-                append(territory)
-                pop()
-                append(" ")
+                if (territory != null) {
+                    pushStyle(MaterialTheme.typography.body2.toSpanStyle())
+                    append(territory)
+                    pop()
+                    append(" ")
+                }
                 pushStyle(codeSpanStyle)
                 append(code)
                 pop()