+++ app/src/test/java/com/mapcode/map/FakeShowMapcodeUseCase.kt
@@ -17,10 +17,14 @@ class FakeShowMapcodeUseCase : ShowMapcodeUseCase {
     var clipboard: String? = null
         private set
 
+    var sharedText: String? = null
+        private set
+
     var hasInternetConnection: Boolean = true
     var currentLocation: Location? = null
 
     val knownLocations: MutableList<FakeLocation> = mutableListOf()
+    val matchingAddresses: MutableMap<String, List<String>> = mutableMapOf()
 
     override fun getMapcodes(lat: Double, long: Double): List<Mapcode> {
         return knownLocations
@@ -80,5 +84,11 @@ class FakeShowMapcodeUseCase : ShowMapcodeUseCase {
         return true
     }
 
-    override fun shareText(text: String, description: String) {}
+    override fun shareText(text: String, description: String) {
+        sharedText = text
+    }
+    
+    override suspend fun getMatchingAddresses(query: String): Result<List<String>> {
+        return success(matchingAddresses[query] ?: emptyList())
+    }
 }