+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## [1.1.0](https://github.com/mapcode-foundation/mapcode-android-app/releases/tag/v1.1.0)
+
+#### 9 August 2022
+
+- [#39](https://github.com/mapcode-foundation/mapcode-android-app/issues/39) Call REST API when querying mapcodes.
+- [#20](https://github.com/mapcode-foundation/mapcode-android-app/issues/20) Show a dropdown that suggests addresses
+  when typing.
+- [#44](https://github.com/mapcode-foundation/mapcode-android-app/issues/44) Don't show a snackbar confirmation when
+  copying to the clipboard on Android 13+ to not conflict with the system's popup.
+- [#43](https://github.com/mapcode-foundation/mapcode-android-app/issues/43) Color the directions button grey.
+- [#40](https://github.com/mapcode-foundation/mapcode-android-app/issues/40) Never show, copy or share the international
+  territory (AAA) for mapcodes.
+- [#41](https://github.com/mapcode-foundation/mapcode-android-app/issues/41) Select all text when tapping latitude and
+  longitude.
+- [#42](https://github.com/mapcode-foundation/mapcode-android-app/issues/42) Copy location coordinates to the clipboard
+  when tapping the latitude and longitude header.
+
 ## [1.0.1](https://github.com/mapcode-foundation/mapcode-android-app/releases/tag/v1.0.1)
 
 #### 5 August 2022